Can now use __getstate__/__setstate__ protocol in anastore
The object creation is more similar to pickle as well: the object is
created by calling on __new__, then updating the instance's __dict__
directly or calling on __setstate__.

Updated signalfilter c++ objects such that they can be
pickled through anastore.

Updated Task because a __setstate__ is now required for correctly
resetting an enum (Level).