checkpath returns a callable handler for opening a track.
Refactored trackio such that the paths can be checked prior to opening a
track. Handler has 2 attributes:

* *path*: is the exact path(s) to a track.
* *handler*: is an object or class which can handle the path

The GR files still require 2 paths, one for the track and another for
the grs. The path to the track can be a regex which will be parsed using
the name of the .cgr file which **must** be available in the directory
with the .gr files.